# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flame Sales & Expense Tracker API",
    "version": "1.0.0",
    "description": "API for managing organizations, projects, expenses, sales, and related resources in the Flame Sales & Expense Tracker."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://flame-expense-tracker.bagumajonah3.workers.dev",
      "description": "Production server (Cloudflare Workers)"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Flame API key in the format flame_ak_<prefix>_<secret>. Include it as an X-Api-Key header when calling these endpoints."
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "employee_name": {
            "type": "string",
            "nullable": true
          },
          "user_role": {
            "type": "string",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "organization_id": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "listUsers",
        "tags": [
          "Users"
        ],
        "summary": "List users in the current organization",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "API key required."
          }
        }
      },
      "delete": {
        "operationId": "deleteCurrentUserAccount",
        "tags": [
          "Users"
        ],
        "summary": "Delete the currently authenticated user account",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Account deleted successfully."
          },
          "401": {
            "description": "API key required."
          }
        }
      },
      "patch": {
        "operationId": "updateCurrentUserProfile",
        "tags": [
          "Users"
        ],
        "summary": "Update profile details for the current user",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "employee_name": {
                    "type": "string",
                    "nullable": true
                  },
                  "phone_number": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully."
          },
          "401": {
            "description": "API key required."
          }
        }
      },
      "put": {
        "operationId": "updateUserRole",
        "tags": [
          "Users"
        ],
        "summary": "Update a user's role (admin only)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "integer"
                  },
                  "role": {
                    "type": "string"
                  }
                },
                "required": [
                  "userId",
                  "role"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User role updated successfully."
          },
          "400": {
            "description": "Missing userId or role."
          },
          "401": {
            "description": "API key required."
          },
          "403": {
            "description": "Forbidden."
          }
        }
      }
    },
...
  }
}